backend/insights_api/pyproject.toml
[project]
name = "insights-api"
version = "0.1.0"
description = "Fastapi api for frontend to query to get product insights"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "pydantic-settings>=2.12.0",
    "valkey[libvalkey]>=6.1.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
]

[tool.pyrefly]
python-version = "3.13"
untyped-def-behavior = "check-and-infer-return-any"
search-path = ["."]

[tool.pyrefly.errors]
unannotated-parameter = "error"
unannotated-return = "error"
unannotated-attribute = "error"
implicit-any = "error"

[dependency-groups]
dev = [
    "pyrefly>=0.48.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
    "testcontainers[postgres]>=4.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
